Todo pasado a to_d

+++ app/helpers/asientos_helper.rb
@@ -4,9 +4,9 @@ module AsientosHelper
     @lista_caja = Hash.new
     asientos.each do |asiento|
       if asiento.tipo_id == 1 then
-        @caja += asiento.monto.to_i
+        @caja += asiento.monto.to_d
       elsif asiento.tipo_id == 2 then
-        @caja -= asiento.monto.to_f
+        @caja -= asiento.monto.to_d
       end
       caja = @caja
       @lista_caja[asiento.id] = caja
